added logger, changed reqs.txt, frontend bug fixed

+++ backend/base/views.py
@@ -16,6 +16,7 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from backend.logger_config import logger
 from backend.settings import MEDIA_URL, MEDIA_ROOT
 from model_integration import process_pdf_2_file, process_text_to_speech
 from .models import Chat, Pdf2FileMessage, Text2Speech
@@ -33,6 +34,7 @@ def user_chats_list_view(request:Request)->Response:
         Response: List of chats in which the user is a member
     """
     username = request.query_params.get('username')
+    logger.info(f"A request was received for a list of chats for the user: {username}")
     # if the user has just registered in the system, then the case User.create() occurs
     user, _ = User.objects.get_or_create(username=username)
     user_chats_list = [
@@ -47,6 +49,7 @@ def user_chats_list_view(request:Request)->Response:
     data = {
       'user_chats': user_chats_list[::-1]
     }
+    logger.info(f"The list of chats for the {username} has been returned")
     return Response(data, status=status.HTTP_200_OK)
 
 
@@ -64,7 +67,12 @@ def get_chat_info_view(request:Request, pk:int)->Response:
         - chat_mode: str
         - messages: list
     """
-    chat = Chat.objects.get(id=pk)
+    logger.info(f"A request was received for information about the chat with the ID: {pk}")
+    try:
+        chat = Chat.objects.get(id=pk)
+    except Chat.DoesNotExist:
+        logger.error(f"The chat was not found with an ID: {pk}")
+        return Response(status=status.HTTP_404_NOT_FOUND)
 
     if chat.mode == 'Extract PDF text':
         messages = Pdf2FileMessage.objects.filter(chat=pk)
@@ -89,6 +97,7 @@ def get_chat_info_view(request:Request, pk:int)->Response:
             }
         for msg in messages]
     else:
+        logger.error("Unsupported chat mode: {chat.mode}")
         return Response(status=status.HTTP_501_NOT_IMPLEMENTED)
     output = {
         "chat_title": chat.title,
@@ -96,6 +105,7 @@ def get_chat_info_view(request:Request, pk:int)->Response:
         "messages": messages,
         "chat_id": chat.id
     }
+    logger.info(f"The chat information with the ID {pk} has been successfully returned")
     return Response(output, status=status.HTTP_200_OK)
 
 
@@ -136,20 +146,28 @@ class ChatEventHandlerView(APIView):
         Returns:
             Response: request status code
         """
+        logger.info(f"A request was received to create a chat with data: {request.data}")
         try:
             request_data = self.PostRequestValidator.model_validate(request.data)
-        except:
+        except Exception as e:
+            logger.error(f"Validation error: {e}")
             return Response(status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
-        user = User.objects.get(username=request_data.username)
+        try:
+            user = User.objects.get(username=request_data.username)
+        except User.DoesNotExist:
+            logger.error(f"The user {request_data.username} was not found")
+            return Response(status=status.HTTP_404_NOT_FOUND)
         new_object = Chat.objects.create(
-            title = request_data.chat_title,
-            mode = request_data.chat_mode,
-            user = user,
+            title=request_data.chat_title,
+            mode=request_data.chat_mode,
+            user=user,
         )
         id = new_object.id
         new_object.href = f'chat/{id}'
         new_object.save()
+
+        logger.info(f"The chat was successfully created with an ID: {id}")
         return Response(status=status.HTTP_201_CREATED)
 
     def put(self, request:Request)->Response:
@@ -161,23 +179,27 @@ class ChatEventHandlerView(APIView):
         Returns:
             Response: request status code
         """
+        logger.info(f"A request was received to change the header of the data chat: {request.data}")
         try:
-            print(request.data)
             request_data = self.PutRequestValidator.model_validate(request.data)
-        except:
+        except Exception as e:
+            logger.error(f"Validation error: {e}")
             return Response(status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
         try:
             chat_instance = Chat.objects.get(id=request_data.chat_id)
-        except:
+        except Chat.DoesNotExist:
+            logger.error(f"The chat {request_data.chat_id} was not found")
             return Response(status=status.HTTP_404_NOT_FOUND)
 
         try:
             chat_instance.title = request_data.new_title
             chat_instance.save()
-        except:
+            logger.info(f"The chat title {request_data.chat_id} has been changed to: {request_data.new_title}")
+        except Exception as e:
+            logger.error(f"Error when saving the chat title: {e}")
             return Response(status=status.HTTP_422_UNPROCESSABLE_ENTITY)
-
+        logger.info("The title has been successfully changed")
         return Response(status=status.HTTP_200_OK)
 
     def delete(self, request:Request)->Response:
@@ -189,13 +211,19 @@ class ChatEventHandlerView(APIView):
         Returns:
             Response: request status code
         """
-
+        logger.info(f"Received a request to delete the chat with data: {request.data}")
         try:
-            chat_instance = Chat.objects.filter(id=request.data['chat_id'])
+            chat_id = request.data['chat_id']
+            chat_instance = Chat.objects.filter(id=chat_id)
+            if not chat_instance.exists():
+                logger.error(f"Chat not found with ID: {chat_id}")
+                return Response(status=status.HTTP_404_NOT_FOUND)
             chat_instance.delete()
-        except:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-
+            logger.info(f"Chat with ID {chat_id} successfully deleted")
+        except Exception as e:
+            logger.error(f"Error while deleting chat: {e}")
+            return Response(status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        logger.info("Chat successfully deleted")
         return Response(status=status.HTTP_200_OK)
 
 
@@ -218,6 +246,7 @@ def create_message_view(request:Request)->Response:
     Returns:
         Response: request status code
     """
+    logger.info(f"Received a request to create a message with data: {request.data}")
     try:
         request_data = MessageValidator(
             chat_id=request.data['chat_id'],
@@ -225,12 +254,14 @@ def create_message_view(request:Request)->Response:
             message=request.data['message'],
             message_type=request.data['message_type'],
             )
-    except:
+    except Exception as e:
+        logger.error(f"Error when saving the chat title: {e}")
         return Response(status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
     try:
         user = User.objects.get(username=request_data.username)
     except:
+        logger.error(f"User not found: {request_data.username}")
         return Response(status=status.HTTP_401_UNAUTHORIZED)
 
     user_chatbot, _ = User.objects.get_or_create(username='chatbot')
@@ -238,6 +269,7 @@ def create_message_view(request:Request)->Response:
     try:
         chat = Chat.objects.get(id=request_data.chat_id)
     except:
+        logger.error(f"Chat not found with ID: {request_data.chat_id}")
         return Response(status=status.HTTP_404_NOT_FOUND)
 
     if chat.mode == 'Extract PDF text':
@@ -247,9 +279,10 @@ def create_message_view(request:Request)->Response:
             fs = FileSystemStorage()
             filename = fs.save(request_data.message.name, request_data.message)
         else:
+            logger.error(f"Invalid document type: {request_data.message_type}")
             return Response(status=status.HTTP_422_UNPROCESSABLE_ENTITY)
         
-        url = os.path.join(os.getenv('SERVER_URL'),os.path.join(MEDIA_URL, filename).lstrip('/')) 
+        url = os.path.join(os.getenv('SERVER_URL'), os.path.join(MEDIA_URL, filename).lstrip('/')) 
 
         Pdf2FileMessage.objects.create(
             chat=chat,
@@ -260,20 +293,19 @@ def create_message_view(request:Request)->Response:
             filesize=request_data.message.size
         )
 
-        is_error = process_pdf_2_file(os.path.join(MEDIA_ROOT, filename), MEDIA_ROOT).returncode
-        if is_error:
-            return Response(status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        process_pdf_2_file(os.path.join(MEDIA_ROOT, filename), MEDIA_ROOT)
 
         markdown_filename = filename.removesuffix('.pdf') + '.mmd'
         # file URL
         url = os.path.join(os.getenv('SERVER_URL'),os.path.join(MEDIA_URL, markdown_filename).lstrip('/')) 
         try:
             filesize = os.path.getsize(os.path.join(MEDIA_ROOT, markdown_filename))
-        except:
+        except Exception as e:
+            logger.warning(f"Failed to get file size: {markdown_filename}. Warning: {e}")
             filesize = 'Unknown'
         Pdf2FileMessage.objects.create(
             chat=chat,
-            user = user_chatbot,
+            user=user_chatbot,
             message=url,
             message_type='file',
             filename=markdown_filename,
@@ -296,5 +328,6 @@ def create_message_view(request:Request)->Response:
             message_type='audio',
         )
     else:
+        logger.error(f"Unsupported chat mode: {chat.mode}")
         return Response(status=status.HTTP_501_NOT_IMPLEMENTED)
     return Response(status=status.HTTP_200_OK)
